feat(server): add event initialize with other client. (#3)
* feat(server): add new type of event return by the server and store current logged user.

* feat(server): missing event.

* style(server): refactor to correspond to coding-style.

+++ libs/protocol/src/server/p_server_listen.c
@@ -9,7 +9,7 @@
 
 static bool read_header(const int fd, p_payload_t *payload)
 {
-    return read(fd, &payload->packet_type, sizeof(uint8_t)) > 0;
+    return read(fd, &payload->packet_type, sizeof(uint16_t)) > 0;
 }
 
 static bool read_network_data(const int fd, p_payload_t *payload,
